Small Changes
Guard added Left Right animation.
ObjectColor Toodoo in previous commit done.
Make my own ProgressBar more softcoded.
Added a AmountOfGold/TotalGold meter up the right corner of the UI.

Had a problem with the Text "hackily" moving when the camera was moving
since the  location of the drawn Text was being rounded. Fixed it by
setting BitmapFont.setUseIntegerPositions(false) (GameScreen Line 48)

The previous problem  I had with LibGDX ProgressBar's Knob was the same
as the problem I had with the Text. Unfortunelately those ProgressBars
didn't have any similar method which could prevent the rounding of the
drawn location of the Knob. I commited a issue to LibGDX asking if they
can add it.