ChasingObject worked on and some methods improved
ChasingGuard(Not decided that it will be a guard looking object) has
been worked on, Now the Chasing "guard" will be following the player
until it is out of sight.

DistanceUtil.isLeft/RightOf method moved to Moving Object method, and
improved.

DistanceUtil.rgbaToRGB method was randomly placed there and so it is
moved to CircleHeist as it will only be a method used to find out about
the RGB of a RGBA color once.

Added optional facing check in the inSight Method(See so the targeter is
facing the target).