Android Fixes, Clean up and more fixes
Optimized sizes of text and buttons for Android.
Hotfixed Android home button making game screen black when clicking on
the game button again.

Cleaned up unnecesary methods and changed their availability(public,
private, protected, default etc).

Changes to some pictures(UI for example).
Fixed the position of some UI objects.

Changed the size of a cacheID to the height of the map which fixes a lag
issue on Android when the map is big(Level 3 etc).

Made a shaperenderer draw a line for the Guards when they are in range
and see you.

Made lots of list hold interfaces instead of the actual class. Added
Textured interface for Textured Objects.

Removed RandomGenerator as nothing used it.

Made Level hold all the data the LevelRenderer should know of.

ObjectColor know handles the Textures.